Finished

+++ src/components/Home/Home.js
@@ -9,27 +9,27 @@ const Home = () => {
             .then(res => res.json())
             .then(data => setUser(data))
     }, [])
-    console.log(user)
+    // console.log(user)
     const handleDelete = id => {
         const process = window.confirm(`Do you want to ser delete?`)
-        console.log(id)
-        const url = `http://localhost:5001/user/${id}`
-        console.log(url)
-        fetch(url, {
-            method: 'DELETE',
-            headers: {
-                "Content-Type": "application/json",
-            },
-            body: JSON.stringify(user)
-        })
-        .then(res=> res.json())
-        .then(data=> {
-            console.log('deleted', user);
-            if (data.deletedCount > 0) {
-                const remaining = user.filter(user => user._id != id);
-                setUser(remaining);
-            }
-        })
+        if (process) {
+            const url = `http://localhost:5001/user/${id}`
+            fetch(url, {
+                method: 'DELETE',
+                headers: {
+                    "Content-Type": "application/json",
+                },
+                body: JSON.stringify(user)
+            })
+                .then(res => res.json())
+                .then(data => {
+                    // console.log('deleted', user);
+                    if (data.deletedCount > 0) {
+                        const remaining = user.filter(user => user._id != id);
+                        setUser(remaining);
+                    }
+                })
+        }
 
 
     }
@@ -39,10 +39,10 @@ const Home = () => {
             <div className="user-full grid gap-4 grid-cols-2 container mx-auto">
                 {
                     user.map(item =>
-                        <div key={item._id}  className=' border border-orange-300 rounded p-4'>
+                        <div key={item._id} className=' border border-orange-300 rounded p-4'>
                             <h1>{item.name}</h1>
                             <p>{item.email}</p>
-                            <button onClick={()=>handleDelete(item._id)} className=' py-2 px-6 bg-red-400 text-white rounded my-3 mr-2'>Delete user</button>
+                            <button onClick={() => handleDelete(item._id)} className=' py-2 px-6 bg-red-400 text-white rounded my-3 mr-2'>Delete user</button>
 
                             <button className=' py-2 px-6 bg-green-600 text-white rounded'>Update user</button>
                         </div>
